AdminPanel v1.3 potwierdzanie/2 bugs fixed

+++ AdminPanel/AdminPanel/Views/Admins.xaml.cs
@@ -69,10 +69,14 @@ namespace AdminPanel.Views
 		}
 		private void LogOut(object sender, RoutedEventArgs e)
 		{
-			Application.Current.Properties["Name"] = "";
-			Application.Current.Properties["ID"] = "";
-			Application.Current.Properties["IfMain"] = "";
-			this.contentControl.Content = new Login(contentControl);
+			MessageBoxResult messageBoxResult = MessageBox.Show("Czy na pewno chcesz się wylogować", "", MessageBoxButton.YesNo, MessageBoxImage.Question);
+			if (messageBoxResult == MessageBoxResult.Yes)
+			{
+				Application.Current.Properties["Name"] = "";
+				Application.Current.Properties["ID"] = "";
+				Application.Current.Properties["IfMain"] = "";
+				this.contentControl.Content = new Login(contentControl);
+			}
 		}
 		private void DataGridView()
 		{
@@ -88,6 +92,7 @@ namespace AdminPanel.Views
 			}
 			conn.Close();
 			DGadmins.ItemsSource = admins;
+			DGadmins.Items.Refresh();
 		}
 		private void Delete(object sender, RoutedEventArgs e)
 		{
@@ -97,11 +102,15 @@ namespace AdminPanel.Views
 				conn.Open();
 				if(Application.Current.Properties["IfMain"].ToString() == "True")
 				{
-					MySqlCommand cmd = conn.CreateCommand();
-					cmd.CommandText = $"DELETE FROM Admins WHERE ID = {id}";
-					cmd.ExecuteNonQuery();
-					conn.Close();
-					DataGridView();
+					MessageBoxResult messageBoxResult = MessageBox.Show("Czy na pewno chcesz usunąć tego administratora", "", MessageBoxButton.YesNo, MessageBoxImage.Question);
+					if (messageBoxResult == MessageBoxResult.Yes)
+					{
+						MySqlCommand cmd = conn.CreateCommand();
+						cmd.CommandText = $"DELETE FROM Admins WHERE ID = {id}";
+						cmd.ExecuteNonQuery();
+						conn.Close();
+						DataGridView();
+					}
 				}
 				else
 				{
