+++ AdminPanel/AdminPanel/Views/Users.xaml.cs
@@ -52,10 +52,14 @@ namespace AdminPanel.Views
 		}
 		private void LogOut(object sender, RoutedEventArgs e)
 		{
-			Application.Current.Properties["Name"] = "";
-			Application.Current.Properties["ID"] = "";
-			Application.Current.Properties["IfMain"] = "";
-			this.contentControl.Content = new Login(contentControl);
+			MessageBoxResult messageBoxResult = MessageBox.Show("Czy na pewno chcesz się wylogować", "", MessageBoxButton.YesNo, MessageBoxImage.Question);
+			if (messageBoxResult == MessageBoxResult.Yes)
+			{
+				Application.Current.Properties["Name"] = "";
+				Application.Current.Properties["ID"] = "";
+				Application.Current.Properties["IfMain"] = "";
+				this.contentControl.Content = new Login(contentControl);
+			}
 		}
 		private void DataGridView()
 		{
@@ -71,6 +75,7 @@ namespace AdminPanel.Views
 			}
 			conn.Close();
 			DGusers.ItemsSource = users;
+			DGusers.Items.Refresh();
 		}
 		private void Delete(object sender, RoutedEventArgs e)
 		{
@@ -80,14 +85,18 @@ namespace AdminPanel.Views
 			conn.Open();
 			if (Application.Current.Properties["IfMain"].ToString() == "True")
 			{
-				MySqlCommand cmd = conn.CreateCommand();
-				cmd.CommandText = $"DELETE FROM benefitstouser WHERE UserID = {id}";
-				cmd.ExecuteNonQuery();
-				cmd = conn.CreateCommand();
-				cmd.CommandText = $"DELETE FROM Users WHERE ID = {id}";
-				cmd.ExecuteNonQuery();
-				conn.Close();
-				DataGridView();
+				MessageBoxResult messageBoxResult = MessageBox.Show("Czy na pewno chcesz usunąć tego użytkownika", "", MessageBoxButton.YesNo, MessageBoxImage.Question);
+				if (messageBoxResult == MessageBoxResult.Yes)
+				{
+					MySqlCommand cmd = conn.CreateCommand();
+					cmd.CommandText = $"DELETE FROM benefitstouser WHERE UserID = {id}";
+					cmd.ExecuteNonQuery();
+					cmd = conn.CreateCommand();
+					cmd.CommandText = $"DELETE FROM Users WHERE ID = {id}";
+					cmd.ExecuteNonQuery();
+					conn.Close();
+					DataGridView();
+				}
 			}
 			else
 			{
